c297: extract DeepLatentsTab + DeepLatentsBody + Stat2 to own file (#441 2.1 + 2.7)
Workspace.tsx 8128 LOC -> 7906 LOC (-222 LOC). Entry chunk 200.70 KB -> 195.85 KB.
#441 P1 2.7 <200 KB target ACHIEVED. 17 tabs extracted + lazy.
DEEP_METHOD_OPTIONS const + CLASS_LABEL_COLORS const moved to new file.
typecheck clean, 44/44 tests pass.
